Fix performance regression
It looks like the performance optimisation introduced in
https://github.com/humio/cloudwatch2humio/pull/9
have been regressed because we are now calling setup() for every
invocation.

This is happening because the `_is_setup` variable being operated on in
helpers.py is different to that in the function files.

To fix this and hopefully prevent this from happening in future we now
scope `_is_setup` to just helpers.py and allow `setup()` to be called on
every invocation.

This has the effect of reducing the average duration from ~900ms -> 25ms
since we aren't creating a new http connection every time.